Add note about using AudioListener2D
There was no mention of the effect of AudioListener2D in this documentation, making it unclear if there was a relationship. The new text is copied and modified from the AudioStreamPlayer3D documentation.

Use 'an' as article before 'AudioListener3D'

(cherry picked from commit a7ebc6fb18bcbcf817895955eed619588b3cd8b8)