Fix gradle warning.
Configuring project ':assetPacks' without an existing directory is
deprecated. The configured projectDirectory
'.../android/java/assetPacks' does not exist, can't be written to or is
not a directory. This behavior has been deprecated. This will fail with
an error in Gradle 9.0. Make sure the project directory exists and can
be written. Consult the upgrading guide for further information:
https://docs.gradle.org/8.11.1/userguide/upgrading_version_8.html#deprecated_missing_project_directory