Deprecated to_json() member methods engine wide. Added as_json() instead.
- Only 3 classes are affected (TOTP, User, HTTPSession).
- This is because gdscript has the built-in to_json method. So to_json()
member methods cannot really be created. (They technically can be done,
but then the global to_json() cannot be called.) As an alternative, a
way to select between a global scope method and a local method could be
created (like ::<method name>() in c++), but because of the way the
engine works, these clashes are extremely rare. (This is the only case I
managed to run into.)