Removed the non-const operator[] from PackedTypedArray.
Due to how this class is going to work, reference based write index
operators cannot be implemented. In order to not trip up anyone else
it's removed. The const version stays, so it can still be indexed, but
set won't work. I.e.: int a = arr[i];: OK arr[i] = a;: Not OK.